Removed browser-telemetry module.
