--- browser-telemetry/browser-telemetry-api/.gitignore
@@ -1,1 +1,0 @@
-/build
