--- browser-telemetry/browser-telemetry-api/src/main/java/com/duckduckgo/browsertelemetry/api/BrowserTelemetryConfig.kt
@@ -1,28 +1,0 @@
-/*
- * Copyright (c) 2023 DuckDuckGo
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.duckduckgo.browsertelemetry.api
-
-/** Public interface for the Browser Telemetry feature */
-interface BrowserTelemetryConfig {
-
-    /**
-     * @return `true` when the remote config has the global "browserTelemetry" feature flag enabled and the
-     * "appLaunch" sub-feature enabled. To be used when we need to temporarily collect extra information on app launch.
-     * @see [BrowserTelemetryFeature]
-     */
-    fun shouldCollectOnAppLaunch(): Boolean
-}
