--- browser-telemetry/browser-telemetry-impl/.gitignore
@@ -1,1 +1,0 @@
-/build
