--- browser-telemetry/browser-telemetry-impl/build.gradle
@@ -1,68 +1,0 @@
-/*
- * Copyright (c) 2023 DuckDuckGo
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-plugins {
-    id 'com.android.library'
-    id 'com.squareup.anvil'
-    id 'kotlin-android'
-}
-
-apply from: "$rootProject.projectDir/gradle/android-library.gradle"
-
-android {
-    anvil {
-        generateDaggerFactories = true // default is false
-    }
-    namespace 'com.duckduckgo.browsertelemetry.impl'
-}
-
-dependencies {
-    implementation project(path: ':app-build-config-api')
-    implementation project(path: ':common-utils')
-    implementation project(path: ':common-ui')
-    implementation project(path: ':di')
-    implementation project(path: ':feature-toggles-api')
-    implementation project(path: ':privacy-config-api')
-    implementation project(path: ':statistics')
-    implementation project(path: ':browser-telemetry-api')
-
-    anvil project(path: ':anvil-compiler')
-    implementation project(path: ':anvil-annotations')
-
-    implementation AndroidX.appCompat
-    implementation Google.android.material
-    implementation AndroidX.constraintLayout
-    implementation AndroidX.core.ktx
-    implementation AndroidX.lifecycle.runtime.ktx
-    implementation AndroidX.lifecycle.viewModelKtx
-    implementation Google.dagger
-    implementation JakeWharton.timber
-    implementation Square.moshi
-    implementation "com.squareup.moshi:moshi-kotlin:_"
-    implementation "com.squareup.moshi:moshi-adapters:_"
-
-    implementation AndroidX.room.ktx
-
-    testImplementation project(path: ':common-test')
-    testImplementation CashApp.turbine
-    testImplementation Testing.junit4
-    testImplementation "org.mockito.kotlin:mockito-kotlin:_"
-    testImplementation (KotlinX.coroutines.test) {
-        // https://github.com/Kotlin/kotlinx.coroutines/issues/2023
-        // conflicts with mockito due to direct inclusion of byte buddy
-        exclude group: "org.jetbrains.kotlinx", module: "kotlinx-coroutines-debug"
-    }
-}
