--- browser-telemetry/browser-telemetry-impl/src/main/java/com/duckduckgo/browsertelemetry/impl/RealBrowserTelemetryConfig.kt
@@ -1,31 +1,0 @@
-/*
- * Copyright (c) 2023 DuckDuckGo
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.duckduckgo.browsertelemetry.impl
-
-import com.duckduckgo.browsertelemetry.api.BrowserTelemetryConfig
-import com.duckduckgo.browsertelemetry.impl.remoteconfig.BrowserTelemetryFeature
-import com.duckduckgo.di.scopes.AppScope
-import com.squareup.anvil.annotations.ContributesBinding
-import javax.inject.Inject
-
-@ContributesBinding(AppScope::class)
-class RealBrowserTelemetryConfig @Inject constructor(
-    private val browserTelemetryFeature: BrowserTelemetryFeature,
-) : BrowserTelemetryConfig {
-
-    override fun shouldCollectOnAppLaunch() = browserTelemetryFeature.self().isEnabled() && browserTelemetryFeature.appLaunch().isEnabled()
-}
