--- browser-telemetry/readme.md
@@ -1,10 +1,0 @@
-# Browser telemetry
-This is a broad term utilized for the temporary collection of additional information (only when specifically enabled) for the purpose of performance analysis on the browser.
-
-One example of this is the collection of the full version of the webview during app launch when the "browserTelemetry" feature and "appLaunch" sub-feature are enabled.
-
-## Who can help you better understand this feature?
-- Ana Capatina
-
-## More information
-- [Asana](https://app.asana.com/0/488551667048375/1205724589991799/f)
